+++ Editor/NativeObjectMultipleEditorsHandlers/RigidbodyMultipleEditors.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 using UnityEditor;
 
-namespace MuffinDev.MultipleEditors.Presets
+namespace MuffinDev.MultipleEditors
 {
 
     /// <summary>