+++ CHANGELOG.md
@@ -4,14 +4,22 @@
 ### Added
 - First release! :)
 
-## [1.1.0] - 2020-10-29
+## [1.1.0] - 2020-11-22
 ### Changed
 - Update to Unity 2019.4+
 - Minor code refactoring
 - Fixed GameObject custom extensions issues
 
-## [1.1.0] - 2020-10-29
+## [1.1.1] - 2020-11-22
 ### Added
 - Online documentation shortcut
 ### Changed
 - Move /_Documentation in /Documentation~
+
+## [1.2.0] - 2020-11-23
+### Added
+- UIElements support (usage of `CreateInspectorGUI()`)
+### Changed
+- Rename Multiple Editors namespace to `MuffinDev.MultipleEditors`
+- Update demo
+- Update documentation